Fixes: edge_src_indices key & cover ax_len=1 case in reduce ops
